-changing price and discount to decimal from double -adding infrastructure project for payment service - implemting client

+++ Repositories/MicroORM/Memento/CodeCatalog.DDD.Memento/CodeCatalog.DDD.Domain/OrderLineFactory.cs
@@ -42,7 +42,7 @@ namespace CodeCatalog.DDD.Domain
                 return false;
             }
 
-            private static double GetDiscountAmount(ProductRequest productRequest)
+            private static decimal GetDiscountAmount(ProductRequest productRequest)
             {
                 return (productRequest.Price * productRequest.Discount) / 100;
             }
