+++ Repositories/MicroORM/Memento/CodeCatalog.DDD.Memento/CodeCatalog.DDD.Domain/OrderLineState.cs
@@ -5,8 +5,8 @@ namespace CodeCatalog.DDD.Domain
     public class OrderLineState
     {
         public ProductId ProductId { get; set; }
-        public double Discount { get; set; }
-        public double Price { get; set; }
+        public decimal Discount { get; set; }
+        public decimal Price { get; set; }
         public uint Quantity { get; set; }
     }
 }
